moved sponsorship section to home folder
