+++ src/lib/sponsors.json
@@ -1,0 +1,34 @@
+[
+  {
+    "name": "sponsor 1",
+    "image": "/placeholder.jpg"
+  },
+  {
+    "name": "sponsor 2",
+    "image": "/placeholder.jpg"
+  },
+  {
+    "name": "sponsor 3",
+    "image": "/placeholder.jpg"
+  },
+  {
+    "name": "sponsor 4",
+    "image": "/placeholder.jpg"
+  },
+  {
+    "name": "sponsor 5",
+    "image": "/placeholder.jpg"
+  },
+  {
+    "name": "sponsor 6",
+    "image": "/placeholder.jpg"
+  },
+  {
+    "name": "sponsor 7",
+    "image": "/placeholder.jpg"
+  },
+  {
+    "name": "sponsor 8",
+    "image": "/placeholder.jpg"
+  }
+]
